moved cli into src/frontend
